refactor: view 패키지를 dto 위로 이동
